# docs/技术架构/openapi/tradehub.openapi.yaml
openapi: 3.1.0
info:
  title: TradeHub API
  version: 1.0.0
  description: |
    TradeHub 第一阶段 API 契约骨架。旧 `/api/v1/*` 接口继续兼容，
    新增能力按 `/api/{domain}/v1/*` 分域推进。
servers:
  - url: http://127.0.0.1:7002
    description: stock-api local
paths:
...
  /api/stock/v1/instruments/search:
    get:
      tags:
        - stock
      summary: 搜索股票标的
      parameters:
        - name: keyword
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/stock/v1/instruments/profile:
    get:
      tags:
        - stock
      summary: 获取股票标的资料
      parameters:
        - name: symbol
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/stock/v1/tasks/ingestion:
    get:
      tags:
        - stock
      summary: 获取股票采集任务
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 创建股票采集任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags:
        - stock
      summary: 删除股票采集任务
      parameters:
        - name: task_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/stock/v1/tasks:
    get:
      tags:
        - stock
      summary: 获取通用股票任务列表或详情
      parameters:
        - name: task_id
          in: query
          required: false
          schema:
            type: string
        - name: task_type
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 创建通用股票任务
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/stock/v1/tasks/status:
    post:
      tags:
        - stock
      summary: 更新通用股票任务状态
      parameters:
        - name: task_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/stock/v1/tasks/logs:
    get:
      tags:
        - stock
      summary: 获取通用股票任务日志
      parameters:
        - name: task_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 追加通用股票任务日志
      parameters:
        - name: task_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/stock/v1/strategies/templates:
    get:
      tags:
        - stock
      summary: 获取策略模板
      parameters:
        - name: enabled_only
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 创建或更新策略模板
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyTemplateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    delete:
      tags:
        - stock
      summary: 删除策略模板
      parameters:
        - name: strategy_id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/stock/v1/strategies/runs:
    get:
      tags:
        - stock
      summary: 获取策略运行记录
      parameters:
        - name: strategy_id
          in: query
          required: false
          schema:
            type: string
        - name: task_id
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 200
            maximum: 1000
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/stock/v1/paper/orders:
    get:
      tags:
        - stock
      summary: 获取模拟交易订单
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      tags:
        - stock
      summary: 提交模拟交易订单
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
  /api/stock/v1/paper/reset:
    post:
      tags:
        - stock
      summary: 重置模拟交易账户
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
...
components:
  schemas:
    ApiMeta:
      type: object
      required:
        - request_id
        - timestamp
      properties:
        request_id:
          type: string
          example: req_20260614_120000_123456
        timestamp:
          type: string
          format: date-time
        version:
          type: string
          example: v1
    ApiErrorDetail:
      type: object
      properties:
        field:
          type: string
        detail:
          type: string
    ApiResponse:
      type: object
      required:
        - success
        - code
        - message
        - data
        - meta
      properties:
        success:
          type: boolean
        code:
          type: string
          example: OK
        message:
          type: string
          example: ok
        data: {}
        meta:
          $ref: "#/components/schemas/ApiMeta"
        error:
          $ref: "#/components/schemas/ApiErrorDetail"
...
    StrategyTemplateInput:
      type: object
      required:
        - name
      properties:
        strategy_id:
          type: string
          example: strategy_tpl_macd_engulfing
        name:
          type: string
          example: MACD 吞噬回测策略
        description:
          type: string
          example: 使用 MACD 红柱和看涨吞噬形态选股，并执行 20 日持有回测
        screening_template_id:
          type: string
          example: screen_tpl_macd_engulfing
        conditions:
          type: object
          additionalProperties: true
        backtest_params:
          type: object
          additionalProperties: true
          example:
            period: 1d
            hold_bars: 20
            lookback: 260
            benchmark_symbol: "000300"
        risk_params:
          type: object
          additionalProperties: true
          example:
            fee_rate: 0.00025
            slippage_rate: 0.0005
            stop_loss: 0.08
            take_profit: 0.15
        enabled:
          type: boolean
          default: true
        created_by:
          type: string
          example: system